Restructure layout tests: rename, split, parameterise
Description
===========
Round-five review feedback. The two test files were misnamed and the
sample-iteration loop in `layout-hierarchy.test.ts` duplicated the one
in `layout-stress.test.ts`. Rename + restructure so each file has one
job:

  - `layout-properties.test.ts` (was layout-hierarchy.test.ts): unit
    tests on structural properties. Every test builds its own inline
    mermaid source. No shared sample fixtures.

  - `layout-samples.test.ts` (was layout-stress.test.ts): sample-driven
    coverage. Parameterised loop over `ALL_SAMPLE_GRAPHS` plus a sanity
    pass over `samples-data.ts` flowcharts.

Changes
=======
- src/__tests__/layout-hierarchy.test.ts → layout-properties.test.ts.
  Strip every test that read from `SAMPLE_GRAPHS` (the pr-98 stress
  case, all nine direction-permutation tests, all four multi-level-
  workflow tests). Strip the canonical-website-samples loop (moved to
  layout-samples.test.ts). The file now contains only inline-source
  property tests: overall graph direction, subgraph direction
  directive respect, three nested-subgraph cross-hier-edge property
  tests, containment, non-overlap.
- src/__tests__/layout-stress.test.ts → layout-samples.test.ts. The
  parameterised loop over `ALL_SAMPLE_GRAPHS` gains an
  `expectedEdgesReachTargets` assertion (covers what `expectEdgeReaches
  Target` in the hierarchy file used to assert). A new describe block,
  `published-gallery flowcharts`, iterates `samples.filter(s =>
  s.category === 'Flowchart')` from samples-data.ts and asserts finite
  positive bounds, no node overlaps, and every leaf inside the graph
  rectangle. File docstring rewritten to explain both sample sets and
  what each asserts.
- src/__tests__/sample-graphs/types.ts: add
  `expectedEdgesReachTargets` to `SampleGraph` so multi-level workflow
  samples can declare their edge-target expectations as data. Drop the
  unused `minGraphWidthOverHeight` field. Reword the file docstring to
  point at `layout-samples.test.ts` (the new name) and clarify that
  every iterated sample gets one `it` per applicable field.
- src/__tests__/sample-graphs/multi-level-workflows.ts: populate
  `expectedNesting` and `expectedEdgesReachTargets` on all four
  multi-* samples so the parameterised loop covers what the deleted
  hierarchy tests used to.
- src/__tests__/sample-graphs/direction-permutations.ts: populate
  `expectedAxisOrder`, `expectedSubgraphAspect`, `expectedNesting`,
  `expectedEdgesReachTargets`, `containment`, and
  `minGraphHeightOverWidth` on all nine perm-* samples. Every
  property the deleted hierarchy tests asserted is now expressed
  as metadata on the corresponding sample.
- src/__tests__/sample-graphs/stress-suite.ts: docstring updated to
  reference `layout-samples.test.ts`.
- tools/sample-comparison/backends/mmc.ts: file docstring now states
  "default dagre renderer" explicitly. The earlier "default renderer"
  phrasing left it ambiguous whether the column was dagre or ELK.
- tools/sample-comparison/README.md: the backend table row for `mmc`
  now reads "mermaid-cli (mmdc) with its default dagre renderer"
  rather than "with ELK enabled". The `--with-mmc` paragraph names
  dagre too. The diff-percentage sentence clarifies that the
  "dimensions" are the rendered SVG's width and height.
- tools/sample-comparison/template-mmc-panel.html: the third column's
  header reads "mermaid-cli (dagre, reference)" rather than
  "mermaid + ELK (reference)".
