+++ src/__tests__/sample-graphs/direction-permutations.ts
@@ -12,6 +12,10 @@ export const DIRECTION_PERMUTATION_SAMPLES: SampleGraph[] = [
     slug: 'perm-lr-with-tb-nested',
     title: 'Permutation: LR root with TB-direction nested subgraph',
     description: 'Outer flow is left-to-right; an inner subgraph declares `direction TB` and so its content stacks vertically, with cross-hierarchy edges entering from the left and leaving on the right of the outer flow.',
+    expectedAxisOrder: [
+      { axis: 'y', items: ['a', 'b', 'c'] },
+    ],
+    expectedSubgraphAspect: [{ subgraph: 'TB Stack', taller: true }],
     source: `graph LR
       subgraph stack [TB Stack]
         direction TB
@@ -24,6 +28,14 @@ export const DIRECTION_PERMUTATION_SAMPLES: SampleGraph[] = [
     slug: 'perm-mixed-siblings',
     title: 'Permutation: TD root with mixed-direction sibling subgraphs',
     description: 'Two sibling subgraphs side-by-side, each declaring its own non-default direction (LR and BT). Each must preserve its own internal direction independently.',
+    expectedAxisOrder: [
+      { axis: 'x', items: ['l1', 'l2', 'l3'] },
+      { axis: 'y', items: ['r3', 'r2', 'r1'] },
+    ],
+    expectedSubgraphAspect: [
+      { subgraph: 'LR pipeline', wider: true },
+      { subgraph: 'BT stack', taller: true },
+    ],
     source: `graph TD
       subgraph leftSide [LR pipeline]
         direction LR
@@ -43,6 +55,11 @@ export const DIRECTION_PERMUTATION_SAMPLES: SampleGraph[] = [
     slug: 'perm-3-level-middle-switch',
     title: 'Permutation: 3-level nesting with one direction switch in the middle',
     description: 'Outer and inner declare TB (matching the root); the middle layer declares LR. Only the middle layer needs SEPARATE_CHILDREN — outer and inner inherit cleanly.',
+    expectedAxisOrder: [
+      { axis: 'x', items: ['a', 'b', 'c'] },
+    ],
+    expectedSubgraphAspect: [{ subgraph: 'LR Middle', wider: true }],
+    expectedEdgesReachTargets: [{ source: 'src', target: 'a' }],
     source: `graph TB
       subgraph outer [TB Outer]
         direction TB
@@ -61,6 +78,14 @@ export const DIRECTION_PERMUTATION_SAMPLES: SampleGraph[] = [
     slug: 'perm-rl-and-bt-siblings',
     title: 'Permutation: sibling subgraphs with RL and BT directions inside a TB parent',
     description: 'Both RL and BT differ from the root TB, so each gets SEPARATE_CHILDREN with FIXED_ORDER ports. Incoming RL ports pin to EAST; incoming BT ports pin to SOUTH (the "start" sides of each direction).',
+    expectedAxisOrder: [
+      { axis: 'x', items: ['rlC', 'rlB', 'rlA'] },
+      { axis: 'y', items: ['btC', 'btB', 'btA'] },
+    ],
+    expectedSubgraphAspect: [
+      { subgraph: 'RL row', wider: true },
+      { subgraph: 'BT stack', taller: true },
+    ],
     source: `graph TB
       subgraph rlGroup [RL row]
         direction RL
@@ -81,6 +106,9 @@ export const DIRECTION_PERMUTATION_SAMPLES: SampleGraph[] = [
     slug: 'perm-rl-in-lr',
     title: 'Permutation: RL-direction nested subgraph reverses flow inside an LR parent',
     description: 'LR and RL flow along the same horizontal axis but in opposite directions. The subgraph still gets SEPARATE_CHILDREN and lays out its content right-to-left.',
+    expectedAxisOrder: [
+      { axis: 'x', items: ['c', 'b', 'a'] },
+    ],
     source: `graph LR
       subgraph reverse [RL Reverse]
         direction RL
@@ -93,6 +121,13 @@ export const DIRECTION_PERMUTATION_SAMPLES: SampleGraph[] = [
     slug: 'perm-4-level-same-direction',
     title: 'Permutation: four-level nesting, all subgraphs matching the root direction',
     description: 'Each subgraph declares the same direction as the root, so none need SEPARATE_CHILDREN — they all flatten via INCLUDE_CHILDREN. A cross-hierarchy edge from the root reaches the deepest leaf naturally without ports.',
+    expectedAxisOrder: [
+      { axis: 'y', items: ['a', 'b', 'c'] },
+    ],
+    expectedNesting: [
+      ['Level 1', 'Level 2', 'Level 3', 'Level 4', 'a'],
+    ],
+    minGraphHeightOverWidth: 1,
     source: `graph TB
       subgraph L1 [Level 1]
         direction TB
@@ -114,6 +149,12 @@ export const DIRECTION_PERMUTATION_SAMPLES: SampleGraph[] = [
     slug: 'perm-alt-lr-tb',
     title: 'Permutation: alternating-direction nesting (LR/LR/TB/LR/TB), only innermost has leaves',
     description: 'Direction swaps at every level except the root match. Intermediate subgraphs hold no leaf nodes — they are pure structure. With no cross-hierarchy edges to route through the multiple SEPARATE_CHILDREN boundaries, the layout still nests cleanly and the innermost direction is preserved.',
+    expectedAxisOrder: [
+      { axis: 'y', items: ['a', 'b', 'c'] },
+    ],
+    expectedNesting: [
+      ['Outer LR', 'Inner TB', 'Deeper LR', 'Deepest TB', 'a'],
+    ],
     source: `graph LR
       subgraph L1 [Outer LR]
         direction LR
@@ -133,6 +174,12 @@ export const DIRECTION_PERMUTATION_SAMPLES: SampleGraph[] = [
     slug: 'perm-alt-tb-lr',
     title: 'Permutation: alternating-direction nesting (TB/TB/LR/TB/LR), only innermost has leaves',
     description: 'Mirror of the previous case starting from TB. Each level alternates direction; only the innermost holds the chain.',
+    expectedAxisOrder: [
+      { axis: 'x', items: ['a', 'b', 'c'] },
+    ],
+    expectedNesting: [
+      ['Outer TB', 'Inner LR', 'Deeper TB', 'Deepest LR', 'a'],
+    ],
     source: `graph TB
       subgraph L1 [Outer TB]
         direction TB
@@ -152,6 +199,11 @@ export const DIRECTION_PERMUTATION_SAMPLES: SampleGraph[] = [
     slug: 'perm-many-cross-hier',
     title: 'Permutation: multiple cross-hierarchy edges into a non-matching direction subgraph',
     description: 'An LR-direction subgraph with two incoming and two outgoing cross-hierarchy edges. With FIXED_ORDER port constraints, all incoming ports pin to the WEST side and all outgoing ports pin to the EAST side.',
+    expectedAxisOrder: [
+      { axis: 'x', items: ['a', 'b', 'c'] },
+    ],
+    expectedSubgraphAspect: [{ subgraph: 'LR Row', wider: true }],
+    containment: { a: 'LR Row', b: 'LR Row', c: 'LR Row' },
     source: `graph TD
       subgraph row [LR Row]
         direction LR
