+++ tools/sample-comparison/template-mmc-panel.html
@@ -1,5 +1,5 @@
     <div class="panel">
-      <div class="panel-head">mermaid + ELK <em>(reference)</em>
+      <div class="panel-head">mermaid-cli <em>(dagre, reference)</em>
         <span class="dim">{{MMC_W}} × {{MMC_H}}</span>
       </div>
       <div class="panel-body">{{MMC_SVG}}</div>
